pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compliance-analysis"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-pinecone>=0.1.0",
    "langgraph>=0.1.0",
    "openai>=1.30.0",
    "pinecone>=5.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "cryptography>=42.0.0",
    "pyyaml>=6.0.1",
    "python-jose[cryptography]>=3.3.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
